+++ src/mailman/app/tests/test_lifecycle.py
@@ -26,6 +26,7 @@ from mailman.app.lifecycle import (
 from mailman.interfaces.address import InvalidEmailAddressError
 from mailman.interfaces.domain import BadDomainSpecificationError
 from mailman.interfaces.listmanager import IListManager
+from mailman.testing.helpers import LogFileMark, configuration
 from mailman.testing.layers import ConfigLayer
 from zope.component import getUtility
 
@@ -45,6 +46,28 @@ class TestLifecycle(unittest.TestCase):
         # raises an exception.
         self.assertRaises(InvalidListNameError,
                           create_list, 'my/list@example.com')
+
+    @configuration('mailman', listname_chars='[a-z0-9-+\]')
+    def test_bad_config_listname_chars(self):
+        mark = LogFileMark('mailman.error')
+        # This list create should succeed but log an error
+        mlist = create_list('test@example.com')
+        # Check the error log.
+        self.assertEqual(
+            mark.readline()[-83:-1],
+            'Bad config.mailman.listname_chars setting: '
+            '[a-z0-9-+\]: '
+            'unterminated character set'
+            )
+        # Remove the list.
+        remove_list(mlist)
+
+    @configuration('mailman', listname_chars='[a-z]')
+    def test_listname_with_minimal_listname_chars(self):
+        # This only allows letters in the listname.  A listname with digits
+        # Raises an exception.
+        self.assertRaises(InvalidListNameError,
+                          create_list, 'list1@example.com')
 
     def test_unregistered_domain(self):
         # Creating a list with an unregistered domain raises an exception.